refactor: walkingDog toggle 로직 walk page에서 useWalkingDogs 로 이동
